fix: move inclusion proof to alpha version
