+++ web/tests/api/clients/inclusion_proof.test.ts
@@ -1,6 +1,6 @@
 import fetchMock from "jest-fetch-mock";
 import { createMocks } from "node-mocks-http";
-import handleInclusionProof from "src/pages/api/v1/clients/inclusion_proof";
+import handleInclusionProof from "src/pages/api/v-alpha/clients/inclusion_proof";
 import { validSequencerInclusionProof } from "../__mocks__/sequencer.mock";
 
 const apiReturnFn = jest.fn();
